Add Endpoint class and refactor type handling in models
- Introduced a new `Endpoint` class to encapsulate API endpoint functionality, including async function handling and Pydantic model validation for parameters and return values.
- Updated the `Response` model to use a generic type variable `G` instead of `T` for better flexibility.
- Refactored imports in `client.py`, `models.py`, and `server.py` to utilize the new `_types.py` module for type definitions, improving code organization and readability.
